Move CrontabSchedule extensions into own file
