+++ NCrontab/CrontabSchedule.cs
@@ -399,122 +399,3 @@ namespace NCrontab
         static Calendar Calendar => CultureInfo.InvariantCulture.Calendar;
     }
 }
-
-namespace NCrontab
-{
-    using System;
-    using System.Collections.Generic;
-    using System.Linq;
-
-    public static class CrontabScheduleExtensions
-    {
-        /// <summary>
-        /// Generates a sequence of unique next occurrences (in order) between
-        /// two dates based on one or more schedules.
-        /// </summary>
-        /// <remarks>
-        /// The <paramref name="baseTime"/> and <paramref name="endTime"/>
-        /// arguments are exclusive.
-        /// </remarks>
-
-        public static IEnumerable<DateTime> GetNextOccurrences(
-            this IEnumerable<CrontabSchedule> schedules,
-            DateTime baseTime, DateTime endTime) =>
-            GetNextOccurrences(schedules, baseTime, endTime, (_, dt) => dt);
-
-        /// <summary>
-        /// Generates a sequence of unique next occurrences (in order) between
-        /// two dates and based on one or more schedules. An additional
-        /// parameter specifies a function that projects the items of the
-        /// resulting sequence where each invocation of the function is given
-        /// the schedule that caused the occurrence and the occurrence itself.
-        /// </summary>
-        /// <remarks>
-        /// The <paramref name="baseTime"/> and <paramref name="endTime"/>
-        /// arguments are exclusive.
-        /// </remarks>
-
-        public static IEnumerable<T> GetNextOccurrences<T>(this IEnumerable<CrontabSchedule> schedules,
-            DateTime baseTime, DateTime endTime,
-            Func<CrontabSchedule, DateTime, T> resultSelector)
-        {
-            if (schedules == null) throw new ArgumentNullException(nameof(schedules));
-            if (resultSelector == null) throw new ArgumentNullException(nameof(resultSelector));
-            return Merge(schedules, s => s.GetNextOccurrences(baseTime, endTime), resultSelector);
-        }
-
-        static IEnumerable<TResult> Merge<T, TSortable, TResult>(
-            IEnumerable<T> sources,
-            Func<T, IEnumerable<TSortable>> sortablesSelector,
-            Func<T, TSortable, TResult> resultSelector)
-            where TSortable : IComparable<TSortable>
-        {
-            var enumerators = new List<KeyValuePair<T, IEnumerator<TSortable>>>();
-
-            try
-            {
-                enumerators.AddRange(from t in sources
-                                     select Pair(t, sortablesSelector(t).GetEnumerator()));
-            }
-            catch
-            {
-                foreach (var e in enumerators)
-                    e.Value.Dispose();
-                throw;
-            }
-
-            try
-            {
-                for (var i = enumerators.Count - 1; i >= 0; i--)
-                {
-                    if (!enumerators[i].Value.MoveNext())
-                    {
-                        enumerators[i].Value.Dispose();
-                        enumerators.RemoveAt(i);
-                    }
-                }
-
-                while (enumerators.Count > 0)
-                {
-                    var i = 0;
-                    var e = enumerators[0];
-
-                    for (var xi = 1; xi < enumerators.Count; xi++)
-                    {
-                        var xe = enumerators[xi];
-                        var comparison = xe.Value.Current.CompareTo(e.Value.Current);
-                        if (comparison < 0)
-                        {
-                            i = xi;
-                            e = xe;
-                        }
-                        else if (comparison == 0)
-                        {
-                            i = xi;
-                            goto skip;
-                        }
-                    }
-
-                    yield return resultSelector(e.Key, e.Value.Current);
-
-                skip:
-
-                    // advance iterator that yielded element, excluding it when consumed
-                    if (!enumerators[i].Value.MoveNext())
-                    {
-                        enumerators[i].Value.Dispose();
-                        enumerators.RemoveAt(i);
-                    }
-                }
-            }
-            finally
-            {
-                foreach (var e in enumerators)
-                    e.Value.Dispose();
-            }
-        }
-
-        static KeyValuePair<TKey, TValue> Pair<TKey, TValue>(TKey key, TValue value) =>
-            new KeyValuePair<TKey, TValue>(key, value);
-    }
-}
